requirements.txt
crewai==0.28.8
crewai_tools==0.1.6
openai==1.13.3
langchain-openai==0.0.5
langchain==0.1.10
langchain-community==0.0.29
docx2txt==0.8
python-dotenv==1.0.1
chromadb==0.4.22
torch = "^1.11.0"
pytest==8.0.0
requests==2.31.0
streamlit==1.36.0
sentence_transformers==3.0.1
faiss-cpu==1.8.0
rank_bm25==0.2.2
openpyxl==3.1.5
docx2txt==0.8
pydantic==2.6.1
python-dotenv==1.0.1
pysqlite3-binary